2010-06-16 21:00 UTC-0500 Vicente Guerra <vicente@guerra.com.mx>   * source/h_grid.prg     ! Deletes new row when edition is cancelled and it's       the first editable column (first grid column can       be readonly).     + Added ( nColumn, aItems ) parameters to evaluate       WHEN column.     + Added ( xValue, nColumn, aItems ) parameters to       evaluate VALID column.     + Added ::OnAppend event.     ! ::OnAppend calls are now using ::DoEvent().   * source/h_xbrowse.prg     - ::OnAppend property is now defined at TGrid.     ! ::OnAppend calls are now using ::DoEvent().   * source/c_controlmisc.c     ! Fixed type cast warning.   * include/oohg.h     ! Patched MSVC warnings for deprecated functions.   * source/h_error.prg     * _OOHG_ErrorMessage() function is now public.

+++ source/c_controlmisc.c
@@ -1,5 +1,5 @@
 /*
- * $Id: c_controlmisc.c,v 1.56 2010-05-15 21:05:05 guerra000 Exp $
+ * $Id: c_controlmisc.c,v 1.57 2010-06-17 02:24:03 guerra000 Exp $
  */
 /*
  * ooHG source code:
@@ -564,8 +564,8 @@ HB_FUNC ( KEYBD_EVENT )
 {
 
 	keybd_event(
-        (WORD) hb_parni(1),                // virtual-key code
-        (WORD) MapVirtualKey( hb_parni(1), 0 ),    // hardware scan code
+        (BYTE) hb_parni(1),                // virtual-key code
+        (BYTE) MapVirtualKey( hb_parni(1), 0 ),    // hardware scan code
 		hb_parl(2) ? KEYEVENTF_KEYUP: 0,	// flags specifying various function options
 		0					// additional data associated with keystroke
 	);
